+++ src/app/interface/categories.ts
@@ -1,8 +1,7 @@
-import { Document } from 'mongoose'
+import { Document, Types } from 'mongoose'
 
 export default interface Icategory extends Document {
     _id: string,
-    user_id: string,
-    category: string,
-    tags: string[],
+    categoryName: string;
+    tags: Types.Array<string>;
 }