+++ src/app/model/categories.ts
@@ -4,19 +4,16 @@ import Icategory from "../interface/categories";
 const categorySchema: Schema<Icategory> = new Schema<Icategory>(
     {
         _id: String,
-        user_id: {
-            type: String,
-            requeried: true
-        },
-        category: {
+        user_id: String,
+        categoryName: {
             type: String,
             requeried: true,
-
         },
         tags: {
             type: [String],
             requeried: true
         }
+
     },
     {
         timestamps: true,