+++ src/function/singJWT.ts
@@ -14,8 +14,10 @@ const signJWT = (user: IUser, callback: (error: Error | null, token: string | nu
 
     try {
         jwt.sign(
-            {
-                username: user.username
+            {   
+                id: user._id,
+                username: user.username,
+                email: user.email
             },
             config.token.secret,
             {