+++ src/routes/routes.ts
@@ -1,6 +1,8 @@
 import express, { Request, Response, Router } from 'express'
 import extractJWT from '../app/middleware/extratJWT'
 import UserController  from '../app/controller/userController'
+import CategoryController from '../app/controller/categoryController'
+import TagsController from '../app/controller/tagsController'
 const router = express.Router()
 
 router.get('/ping', (req: Request ,res: Response) => {
@@ -9,10 +11,18 @@ router.get('/ping', (req: Request ,res: Response) => {
 
 router.post('/create_user',UserController.create)
 router.post('/login_user', UserController.login)
-// router.post('/delete_user', UserController.delete)
-router.get('/validate_user', extractJWT,UserController.validateToken)
-router.post('/create_user', UserController.create)
-router.post('/create_user', UserController.create)
-router.post('/create_user', UserController.create)
 
+router.use(extractJWT)
+
+/**
+ * category routes
+ */
+router.post('/create_category', CategoryController.createCategory)
+router.delete('/delete_category/:category', CategoryController.deleteCategory)
+router.get('/get_categories', CategoryController.getCategories)
+
+/**
+ * tags routes
+ */
+router.post('/create_tag', TagsController.create)
 export default router