+++ changelog.md
@@ -5,7 +5,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+* tags controller
+- delete
+- change
+- get all
+## [0.0.2] - 2017-06-20
+### Added
+- controller and routes to tags and category
+### Changed
+- response from token when calling them
+- relation of tables user and categories
+### Removed
+- route and method validator
+## [Unreleased]
+* tags controller
+- create
+- delete
+- change
+- get all
 ## [0.0.1] - 2017-06-20
 ### Added
 - authenticate token